Fail on missing bundle paths
If a bundle path is specified that does not exists, it was previously
silently ignored. Surprising behavior will thus occur when making a
typo or specifying a path outside the Nix store.
Fail in this case by raising an exception, so that the user is notified
of their mistake and can fix it.