fix: repair `shell.nix` up to modern Nixpkgs
The `shell.nix` was pinning various things like Python 3.9, specific versions of dependencies, had
a reference to Python 2, was using a removed package / moved package (pre-commit) and was overriding
`reuse` to inject a specific set of packages.

All of this is cleaned up and the shell is now using Python 3.12 if you are using a recent Nixpkgs, at the time
of writing.

In addition, `pytest` and `pre-commit` are still behaving as expected.

Signed-off-by: Ryan Lahfa <ryan@lahfa.xyz>